DeleteCardModal moved into CardsView directory
