+++ spaced-repetition-front-end/src/components/CardsView/DeleteCardModal.js
